+++ cardgamesgui/src/main/java/cards/gui/dialogs/EditorTarot.java
@@ -25,6 +25,7 @@ import cards.tarot.GameTarot;
 import cards.tarot.HandTarot;
 import cards.tarot.sml.DocumentWriterTarotUtil;
 import code.gui.ConfirmDialog;
+import code.gui.ConstFiles;
 import code.gui.FileSaveDialog;
 import code.gui.LabelButton;
 import code.gui.Panel;
@@ -34,7 +35,6 @@ import code.util.CustList;
 import code.util.EqList;
 import code.util.Numbers;
 import code.util.StringList;
-import code.util.consts.ConstFiles;
 
 public final class EditorTarot extends DialogTarot implements SetterSelectedCardList {
     private static final String DIALOG_ACCESS = "cards.gui.dialogs.editortarot";
@@ -311,7 +311,7 @@ public final class EditorTarot extends DialogTarot implements SetterSelectedCard
         byte donneur_ = (byte) liste.getSelectedIndex();
         if (donneur_ == nombreDeJoueurs_) {
 //            donneur_=(byte)Math.floor(nombreDeJoueurs_*MonteCarlo.randomDouble());
-            donneur_=(byte)AbMonteCarlo.randomInt(nombreDeJoueurs_);
+            donneur_=(byte)AbMonteCarlo.randomLong(nombreDeJoueurs_);
         }
         DealTarot donne_=new DealTarot(mains_,donneur_);
         partie = new GameTarot(GameType.EDIT,donne_,getReglesTarot());
