+++ tarot/src/main/java/cards/tarot/HandTarot.java
@@ -213,7 +213,7 @@ public final class HandTarot implements Iterable<CardTarot>, Equallable<HandTaro
     @return la carte aleatoire choisie*/
     CardTarot tirerUneCarteAleatoire() {
 //        return jouer((int)Math.floor(total()*MonteCarlo.randomDouble()));
-        return jouer(AbMonteCarlo.randomInt(total()));
+        return jouer((int) AbMonteCarlo.randomLong(total()));
         //0<=total()*Math.random()<total()
         //Donc 0<=Math.floor(total()*Math.random())<Math.floor(total())=total()
     }
